1OX22CS116-2FA-Project/requirements.txt
PyQt5>=5.15.9
Pillow>=10.0.0
stegano>=0.11.2
cryptography>=42.0.0